Add infrastructure for handling different trace versions
Split trace parser to specific version handling
