+++ source/octf/trace/parser/CMakeLists.txt
@@ -1,3 +1,5 @@
+add_subdirectory(v4)
+
 target_sources(octf
 PRIVATE
     ${CMAKE_CURRENT_LIST_DIR}/IoTraceEventHandlerCsvPrinter.h
