+++ source/octf/trace/parser/v4/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(octf
+PRIVATE
+    ${CMAKE_CURRENT_LIST_DIR}/ParsedIoTraceEventHandler.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/ParsedIoTraceEventHandler.h
+)
